završeno dodavanje AdSistemAlarmi, zajedničkih interfejsa i abstract klasa

+++ service/src/main/java/rs/prati/service/repository/AaSistemRepository.java
@@ -4,7 +4,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import rs.prati.core.model.AaSistem;
 
+/**
+ * Репозиторијум за приступ ентитету AaSistem.
+ */
 @Repository
-public interface AaSistemRepository extends JpaRepository<AaSistem, Long> {
-	
+public interface AaSistemRepository extends JpaRepository<AaSistem, Integer> {
+    // Нема додатних метода; користимо findById(1) за приступ систему
 }
